camera2: Ask for PREVENT_POWER_KEY, not PREVENT_SYSTEM_KEYS permission
Prior to this change, enabling Power Shutter in advanced options
causes the camera to crash and then to subsequently crash each time
it is opened until you clear its settings.  The crash is that it does
not have the required permission PREVENT_POWER_KEY.

Looking through git it seems that we switched from PREVENT_POWER_KEY to
PREVENT_SYSTEM_KEYS in

commit 49896fc2449b51033e9631bdcc186aeaf71701ff
Author: Adnan Begovic <adnan@cyngn.com>
Date:   Fri May 20 13:01:40 2016 -0700

    Camera2: Use new window manager flag for prevent system keys.

and then partially reverted it in:

commit 47fb21ee41b78e8c1d5e353714289650c5a18c29
Author: Scott Mertz <scott@cyngn.com>
Date:   Tue May 31 17:48:09 2016 -0700

    Camera2: only override power key

    Partial revert "Camera2: Use new window manager flag for prevent system keys."

    CYNGNOS-2958
    Change-Id: I3dc95f8b150e673bc9cada437e121ec3ceb46b84

The revert did not revert back to the required permission.

Change-Id: I09c52ac9e63ed725a65ed87e8cea99d942a97846
Signed-off-by: CheckYourScreen <nimitmehta95@gmail.com>